SimpleQueue and Empty live in _queue
