+++ stdlib/queue.pyi
@@ -1,4 +1,5 @@
 import sys
+from _queue import Empty as Empty, SimpleQueue as SimpleQueue
 from threading import Condition, Lock
 from typing import Any, Generic, TypeVar
 
@@ -11,7 +12,6 @@ if sys.version_info >= (3, 13):
 
 _T = TypeVar("_T")
 
-class Empty(Exception): ...
 class Full(Exception): ...
 
 if sys.version_info >= (3, 13):
@@ -55,14 +55,3 @@ class PriorityQueue(Queue[_T]):
 
 class LifoQueue(Queue[_T]):
     queue: list[_T]
-
-class SimpleQueue(Generic[_T]):
-    def __init__(self) -> None: ...
-    def empty(self) -> bool: ...
-    def get(self, block: bool = True, timeout: float | None = None) -> _T: ...
-    def get_nowait(self) -> _T: ...
-    def put(self, item: _T, block: bool = True, timeout: float | None = None) -> None: ...
-    def put_nowait(self, item: _T) -> None: ...
-    def qsize(self) -> int: ...
-    if sys.version_info >= (3, 9):
-        def __class_getitem__(cls, item: Any, /) -> GenericAlias: ...
